+++ app/controllers/share_file_controller.rb
@@ -40,6 +40,14 @@ class ShareFileController < ApplicationController
     @share_file.publication_type = params[:publication_type]
     @share_file.publication_symbols_value = params[:publication_type]=='protected' ? params[:publication_symbols_value] : ""
 
+    # 所有者がマイユーザ OR マイグループ
+    unless login_user_symbols.include? params[:owner_symbol]
+      @categories_hash = ShareFile.get_tags_hash(params[:owner_symbol])
+      flash.now[:warning] = "この操作は、許可されていません。"
+      render :action => "new"
+      return
+    end
+    
     unless valid_upload_files? params[:file]
       @categories_hash = ShareFile.get_tags_hash(params[:owner_symbol])
       flash.now[:warning] = "ファイルサイズが０、もしくはサイズが大きすぎるファイルがあります。"
@@ -52,7 +60,9 @@ class ShareFileController < ApplicationController
       @share_file.file_name = file.original_filename
       @share_file.content_type = file.content_type.chomp
 
-      if not (verify_message = verify_file_size(file)) and @share_file.save
+      if not (verify_message = verify_extension_share_file(@share_file)) and 
+        not (verify_message = verify_file_size(file)) and 
+        @share_file.save
         target_symbols = analyze_param_publication_type
         target_symbols.each do |target_symbol|
           @share_file.share_file_publications.create(:symbol => target_symbol)
@@ -92,6 +102,11 @@ class ShareFileController < ApplicationController
       return
     end
 
+    unless authorize_to_save_share_file? @share_file 
+      render_window_close
+      return
+    end
+    
     params[:publication_symbols_value] = ""
     if @share_file.public?
       params[:publication_type] = "public"
@@ -113,6 +128,13 @@ class ShareFileController < ApplicationController
     update_params = params[:share_file]
     update_params[:publication_type] = params[:publication_type]
     update_params[:publication_symbols_value] = params[:publication_type]=='protected' ? params[:publication_symbols_value] : ""
+
+    unless authorize_to_save_share_file? @share_file
+      @categories_hash = ShareFile.get_tags_hash(@share_file.owner_symbol)
+      flash.now[:warning] = "この操作は、許可されていません。"
+      render :action => :edit
+      return
+    end
 
     if @share_file.update_attributes(update_params)
       @share_file.share_file_publications.clear
@@ -138,6 +160,12 @@ class ShareFileController < ApplicationController
       return false
     end
 
+    unless authorize_to_save_share_file? share_file
+      flash[:warning] = "この操作は、許可されていません。"
+      redirect_to :controller => share_file.owner_symbol_type, :action => share_file.owner_symbol_id, :id => 'share_file'
+      return
+    end
+    
     if share_file.destroy
       flash[:notice] = "ファイルの削除に成功しました。"
     else
@@ -214,6 +242,11 @@ class ShareFileController < ApplicationController
       return false
     end
 
+    unless authorize_to_save_share_file? share_file
+      send_data(" ", :filename => "dummy.csv", :type => 'application/x-csv', :disposition => 'attachment')
+      return
+    end
+
     csv_text, file_name = share_file.get_accesses_as_csv
     send_data(csv_text, :filename => nkf_file_name(file_name), :type => 'application/x-csv', :disposition => 'attachment')
   end
@@ -221,6 +254,11 @@ class ShareFileController < ApplicationController
   def clear_download_history
     unless share_file = ShareFile.find(:first, :conditions => ["id = ?", params[:id]])
       return false
+    end
+
+    unless authorize_to_save_share_file? share_file
+      render :nothing => true
+      return
     end
 
     share_file.share_file_accesses.clear
@@ -267,10 +305,26 @@ private
     return nil
   end
 
+  def verify_extension_share_file file
+      unless verify_extension? file.file_name, file.content_type
+        return "この形式のファイルは、アップロードできません。" 
+      end
+      return nil
+  end
+
+
   def nkf_file_name(file_name)
     agent = request.cgi.env_table["HTTP_USER_AGENT"]
     return  NKF::nkf('-Ws', file_name) if agent.include?("MSIE") and not agent.include?("Opera")
     return file_name
+  end
+
+  # 権限チェック 
+  # ・所有者がマイユーザ
+  # ・所有者がマイグループ  AND 作成者がマイユーザ 
+  def authorize_to_save_share_file? share_file
+    session[:user_symbol] == share_file.owner_symbol || 
+      (login_user_groups.include?(share_file.owner_symbol) && (session[:user_id] == share_file.user_id))
   end
 
 end
